+++ BeerApp.DataAccess/Models/UserFavoriteBeer.cs
@@ -5,10 +5,10 @@ namespace BeerApp.DataAccess.Models
 	[Table("Favorites")]
     public class UserFavoriteBeer
     {
-		public long UserId { get; set; }
+		public int UserId { get; set; }
 		public User User { get; set; }
 
-		public long BeerId { get; set; }
+		public int BeerId { get; set; }
 		public Beer Beer { get; set; }
 	}
 }
