+++ BeerApp.DataAccess/Repositories/interfaces/IFavoritesRepository.cs
@@ -8,7 +8,7 @@ namespace BeerApp.DataAccess.Repositories
     {
 	    Task<UserFavoriteBeer> AddAsync(UserFavoriteBeer favorite);
 	    Task<UserFavoriteBeer> RemoveAsync(UserFavoriteBeer favorite);
-	    Task<IReadOnlyList<Beer>> GetAllAsync(long userId);
-	    Task<UserFavoriteBeer> FindAsync(long userId, long beerId);
+	    Task<IReadOnlyList<Beer>> GetAllAsync(int userId);
+	    Task<UserFavoriteBeer> FindAsync(int userId, int beerId);
     }
 }
