+++ BeerApp.Web/Models/Search/SearchParams.cs
@@ -2,13 +2,13 @@
 {
 	public class SearchParams
 	{
-		public long? PerPage { get; } = 9;
-		public long? Page { get; set; } = 1;
+		public int? PerPage { get; } = 9;
+		public int? Page { get; set; } = 1;
 
 		public string BeerName { get; set; }
 		
-		public int? Abv { get; set; }
-		public int? Ebc { get; set; }
-		public int? Ibu { get; set; }
+		public float? Abv { get; set; }
+		public float? Ebc { get; set; }
+		public float? Ibu { get; set; }
 	}
 }
